+++ CHANGELOG.md
@@ -31,6 +31,8 @@ CHANGELOG
   will be used for any child resource of a component or stack.
   [#3174](https://github.com/pulumi/pulumi/pull/3174)
 
+- Add resource transformations support in Python. [#3319](https://github.com/pulumi/pulumi/pull/3319)
+
 ## 1.2.0 (2019-09-26)
 
 - Support emitting high-level execution trace data to a file and add a debug-only command to view trace data.